+++ src/media_enrichment/__init__.py
@@ -1,7 +1,7 @@
 """media-enrichment Skill package.
 
-Version: 0.1.0-dev13
+Version: 0.1.0-dev14
 """
 
-__version__ = "0.1.0-dev13"
+__version__ = "0.1.0-dev14"
 SKILL_NAME = "media-enrichment"